feat: test result 라우트 이동시 로딩기능 추가
