+++ src/App.tsx
@@ -1,16 +1,8 @@
 import { Routes, Route } from "react-router-dom";
-import { Home, TestPage, SearchPage, Detail } from "./pages";
+import { Home, TestPage, TestResultPage, SearchPage, Detail } from "./pages";
 import { DataLoad, Layout, NotFound } from "./components";
-import { Suspense, lazy } from "react";
 
 export default function App(): JSX.Element {
-  const TestResultPage = lazy(() => {
-    const promise1 = import("./pages/TestResultPage");
-    const promise2 = new Promise((resolve) => setTimeout(resolve, 0));
-
-    return Promise.all([promise1, promise2]).then(([moduleExports]) => moduleExports);
-  });
-
   return (
     <Routes>
       <Route
